Move Status to patches & replace Procedure status
